Time: 205 ms (19.2%), Space: 55 MB (18.8%) - LeetHub

+++ 2610-closest-prime-numbers-in-range/2610-closest-prime-numbers-in-range.java
@@ -6,8 +6,8 @@ class Solution {
         prime[1]=false;
         for(int i=2;i*i<=right;i++){
             if(prime[i]){
-                for(long j=i*i;j<=right;j+=i){
-                    prime[(int)j]=false;
+                for(int j=i*i;j<=right;j+=i){
+                    prime[j]=false;
                 }
             }
         }
